refactor: delete team confirmation dialog
